simplify: user control functions in Map.js
